backend/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy==2.0.35
alembic==1.13.3

# SQLite doesn't need extra driver, PostgreSQL needs:
# psycopg2-binary==2.9.9

# Anthropic API
anthropic==0.39.0

# File handling
aiofiles==24.1.0

# Authentication
PyJWT==2.9.0
cryptography==43.0.1
httpx==0.27.2

# Utilities
python-dotenv==1.0.1
pydantic-settings==2.5.2
